Implemented list request
-Replica can now recieve and process "list" requests on the recovery
port
-proxy can recieve and forward "list" requests to a replica
-proxy will try a different replica if the first fails
